gauleses/requirements.txt
django-bootstrap-modal-forms==2.2.0
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2.5
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
Markdown==3.3.4
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.2.0
python-decouple==3.6
unidecode==1.3.4
numpy
pandas
openpyxl==3.0.9
boto3==1.18.52
django-storages==1.13.1